+++ packages/react/src/index.ts
@@ -1,3 +1,3 @@
-export * as Forms from "./exports/forms";
-export { useOptionalExternalState } from "./hooks/useOptionalExternalState";
-export * as Conditional from "./components/Conditional";
+export * as Forms from "./exports/forms.js";
+export { useOptionalExternalState } from "./hooks/useOptionalExternalState.js";
+export * as Conditional from "./components/Conditional.js";