+++ packages/react/vitest.config.ts
@@ -20,7 +20,7 @@ export default defineProject({
 							process.env.LOCALHOST_TLS_KEY_FILE ??
 								`${userHomeDir}/.certs/key.pem`,
 						),
-				  }
+					}
 				: undefined,
 	},
 	test: {