+++ packages/coding-standards/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @atomicsmash/coding-standards
 
+## 9.0.0-beta.1
+
+### Patch Changes
+
+- Fix builds
+
+- Updated dependencies []:
+  - @atomicsmash/browserslist-config@9.0.0-beta.1
+  - @atomicsmash/eslint-config@9.0.0-beta.1
+
+## 9.0.0-beta.0
+
+### Major Changes
+
+- [`38d8dac`](https://github.com/AtomicSmash/packages/commit/38d8dac02de80f2a0dfa673194a401e2a69aed55) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Update dependencies
+
+- [`38d8dac`](https://github.com/AtomicSmash/packages/commit/38d8dac02de80f2a0dfa673194a401e2a69aed55) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Update typescript config files
+
+### Patch Changes
+
+- Updated dependencies [[`38d8dac`](https://github.com/AtomicSmash/packages/commit/38d8dac02de80f2a0dfa673194a401e2a69aed55)]:
+  - @atomicsmash/eslint-config@9.0.0-beta.0
+  - @atomicsmash/browserslist-config@9.0.0-beta.0
+
 ## 8.0.0
 
 ### Major Changes