+++ packages/dev-tools-helpers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atomicsmash/dev-tools-helpers
 
+## 1.2.0-beta.1
+
+### Patch Changes
+
+- Fix builds
+
+## 1.2.0-beta.0
+
+### Minor Changes
+
+- [`38d8dac`](https://github.com/AtomicSmash/packages/commit/38d8dac02de80f2a0dfa673194a401e2a69aed55) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Update dependencies
+
 ## 1.1.1
 
 ### Patch Changes